Extract service class for making the code more testable. Add test cases for the null pointer fix issue.
